+++ cmp/markdown/Markdown.ts
@@ -7,11 +7,11 @@
 import {hoistCmp, HoistProps} from '@xh/hoist/core';
 import {reactMarkdown} from '@xh/hoist/kit/react-markdown';
 import remarkBreaks from 'remark-breaks';
+import remarkGfm from 'remark-gfm';
+import {PluggableList} from 'unified/lib';
 
 interface MarkdownProps extends HoistProps {
-    /**
-     * Markdown formatted string to render.
-     */
+    /** Markdown formatted string to render. */
     content: string;
 
     /** True (default) to render new lines with <br/> tags. */
@@ -20,13 +20,18 @@ interface MarkdownProps extends HoistProps {
 
 /**
  * Render Markdown formatted strings as HTML (e.g. **foo** becomes <strong>foo</strong>).
+ *
+ * Note that the remark-gfm plugin is included by default to support GitHub Flavored Markdown,
+ * a superset of the CommonMark specification. See https://github.github.com/gfm/ for details.
  */
 export const [Markdown, markdown] = hoistCmp.withFactory<MarkdownProps>({
     displayName: 'Markdown',
     render({content, lineBreaks = true}) {
+        const remarkPlugins: PluggableList = [remarkGfm];
+        if (lineBreaks) remarkPlugins.push(remarkBreaks);
         return reactMarkdown({
             item: content,
-            remarkPlugins: lineBreaks ? [remarkBreaks] : null
+            remarkPlugins
         });
     }
 });